+++ src/room_layout.py
@@ -1,0 +1,11 @@
+import bpy
+
+
+def _register():
+    # bpy.utils.register_class(__name__)
+    pass
+
+
+def _unregister():
+    # bpy.utils.unregister_class(__name__)
+    pass
